# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: DigicoreOS Platform API
  version: 1.0.0
  description: |
    Public HTTP contract for the AI-first Rust SaaS platform — six bounded
    contexts (Auth, ERP, CRM, HRM, Reporting, AI) fronted by a single edge under
    `/api/v1/`. Monetary amounts are integers in the currency's minor units
    (e.g. cents) paired with an ISO-4217 `currency`. Identifiers are UUIDs;
    timestamps are RFC 3339. List endpoints accept `page`/`page_size` and return a
    paginated envelope.
  license:
    name: Proprietary
servers:
  - url: https://api.digicore.example.com
    description: Production edge (replace host to match deploy/k8s/50-ingress.yaml)
  - url: http://localhost:8080
    description: Local development (per-service ports 8081-8086 when not behind the edge)
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/trade-export/shipments:
    get:
      tags: [TradeExport]
      summary: List export shipments
      operationId: listShipments
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: A page of shipments
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ShipmentPage' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [TradeExport]
      summary: Create an export shipment
      operationId: createShipment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Shipment' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Shipment' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/trade-export/shipments/{shipment_id}:
    parameters:
      - name: shipment_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [TradeExport]
      summary: Get an export shipment
      operationId: getShipment
      responses:
        '200':
          description: The shipment
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Shipment' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/trade-export/shipments/{shipment_id}/book:
    parameters:
      - name: shipment_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [TradeExport]
      summary: Book a draft shipment (DRAFT→BOOKED, emits ShipmentBooked)
      operationId: bookShipment
      responses:
        '200':
          description: Booked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Shipment' }
        '409': { $ref: '#/components/responses/Conflict' }
        default: { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 access token from /api/v1/auth/login, verified per service.
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
...
  responses:
    Error:
      description: Unexpected error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Invalid state transition for the resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: Machine-readable error code, example: validation_error }
        message: { type: string, description: Human-readable message }
        request_id: { type: string, description: Correlates with server logs/traces }
        details:
          type: array
          description: Optional field-level problems.
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    PageMeta:
      type: object
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }

    # ---- Auth ----
...
    ShipmentStatus:
      type: string
      enum: [DRAFT, BOOKED, DISPATCHED, CANCELLED]
    Shipment:
      type: object
      required: [destination_country, incoterm]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        tenant_id: { type: string, readOnly: true }
        order_id:
          type: [string, "null"]
          description: ERP order this shipment fulfils (null for a manual shipment).
        reference: { type: string, readOnly: true, description: "e.g. EXP-1A2B3C4D" }
        destination_country: { type: string, description: ISO-3166 alpha-2, example: VN }
        incoterm: { type: string, example: FOB }
        status: { $ref: '#/components/schemas/ShipmentStatus' }
        created_at: { type: string, format: date-time, readOnly: true }
    ShipmentPage:
      type: object
      allOf: [{ $ref: '#/components/schemas/PageMeta' }]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Shipment' } }